[refactor] ファイル構造の整理とサフィックス削除によるコーディング規約準拠
- resolverをsrc/presentation/resolver/に移動
- ファイル名からsuffix削除: .resolver.ts, .dto.ts, .usecase.ts
- import文を全て更新
- docs/coding-standards.md と docs/github-issue-workflow.md にファイル命名規則を追加
