+++ src/modules/user/user.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserController } from '../../presentation/user.controller';
-import { UserUsecase } from '../../usecase/user.usecase';
+import { UserUsecase } from '../../usecase/user';
 import { User } from '../../domain/entity/user';
-import { UserResolver } from '../../graphql/resolvers/user.resolver';
+import { UserResolver } from '../../presentation/resolver/user';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
